Style Main section + Add profile pic
Style the Intro, Portfolio and Contact sections.

For Intro, add profile picture + some descriptive text
about the author of the portfolio.

For Portfolio, add preview image/GIF which, when clicked, opens
a modal that displays the image/GIF in more details. Add external
links to the Github repositories of each project.

In Contact, add text that points out the author's email address.